Edit SubProject now implemented, with corresponding mapping and methods implemented.
Had several issues trying to implement editable deadlines for subprojects, we'll revist this issue later. For now ittles of subprojects can be edited, so using an alternative String for deadline attribute should be simple, as well as including a subproject description.

Note: Several debugging line are still included for better handling the areas I'll revisit.

resolves #49